* refactored and simplified transfer api * use more GlazedLists stuff (EventList, AutoCompleteSupport) and remove obsolete classes (SimpleListModel, TextCompletion) * don't use SearchResultCache in EpisodeListClient (was only done for better ui interactions) * removed caching from ResourceManager * some improvements based on FindBugs warnings * use args4j for improved argument parsing * updated ant build script * more general MessageBus/Handler (use Object as message type instead of string) * ChecksumComputationService is not a singleton anymore * TemporaryFolder is always recreated if it is deleted by the user, or another instance shutting down * Notifications flicker less when one window is removed and the others are layouted * lots of other refactoring

+++ source/net/sourceforge/tuned/FileUtil.java
@@ -16,9 +16,9 @@ public class FileUtil {
 	
 	public static String formatSize(long size) {
 		if (size >= MEGA)
-			return String.format("%d MB", (double) size / MEGA);
+			return String.format("%d MB", size / MEGA);
 		else if (size >= KILO)
-			return String.format("%d KB", (double) size / KILO);
+			return String.format("%d KB", size / KILO);
 		else
 			return String.format("%d Byte", size);
 	}
